apps/services/tax-engine/pyproject.toml
[tool.poetry]
name = "apgms-tax-engine"
version = "0.1.0"
description = "APGMS Tax Engine (P11)"
authors = ["you"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.2"
uvicorn = { extras = ["standard"], version = "^0.30.6" }
pydantic = "^2.9.2"
sqlalchemy = "^2.0.36"
psycopg = { version = "^3.2.3", extras = ["binary"] }
alembic = "^1.13.3"
nats-py = "^2.11.0"
prometheus-client = "^0.21.0"
orjson = "^3.10.7"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
